fix test location
